model/requirements.txt
Flask==3.0.3
huggingface-hub==0.25.2
Pillow==10.4.0
python-dotenv==1.0.1
